+++ vJoyMapping/Thrustmaster/Warthog/Throttle/Mapping/TmThrottleLandedStateHandler.cs
@@ -19,14 +19,14 @@ namespace vJoyMapping.Thrustmaster.Warthog.Throttle.Mapping
             {
                 // Set the left throttle to the middle so we don't move forward or backward when the throttle is parked
                 controller.SharedState.LeftThrottleEnabled = false;
-                controller.SetJoystickAxis(16 * 1024, vJoy.Wrapper.Axis.HID_USAGE_RZ, vJoyTypes.Throttle);
+                controller.SetJoystickAxis(16 * 1024, (int)vJoy.Wrapper.Axis.HID_USAGE_RZ, vJoyTypes.Throttle);
                 controller.Logger.LogDebug("Left Throttle Parked");
             }
             else if (Reactive.ButtonReleased(value, leftThrottleParked))
             {
                 // Left throttle is set to move forward and backward from the center 
                 controller.SharedState.LeftThrottleEnabled = true;
-                controller.SetJoystickAxis(32 * 1024, vJoy.Wrapper.Axis.HID_USAGE_RZ, vJoyTypes.Throttle);
+                controller.SetJoystickAxis(32 * 1024, (int)vJoy.Wrapper.Axis.HID_USAGE_RZ, vJoyTypes.Throttle);
                 controller.Logger.LogDebug("Left Throttle Activated");
             }
         }
